requirements.txt
pytube
moviepy
openai
python-dotenv
yt_dlp